+++ src/rcpp_calculate_ci.h
@@ -7,6 +7,6 @@ using namespace Rcpp;
 NumericVector rcpp_calculate_ci(NumericMatrix matrix,
                                 double alpha,
                                 double beta,
-                                double max_dist);
+                                int max_dist);
 
 #endif
